Use vitepress-md-var package from npm in place of direct embedding
